fix: block.timestmap to uint248

+++ test/integration/perps/order-module/createOrder/createOrder.t.sol
@@ -32,7 +32,8 @@ contract CreateOrder_Integration_Test is Base_Integration_Shared_Test {
             acceptablePrice: uint128(MOCK_ETH_USD_PRICE),
             orderType: Order.OrderType.MARKET
         });
-        Order.Data memory expectedOrder = Order.Data({ id: 0, payload: payload, settlementTimestamp: block.timestamp });
+        Order.Data memory expectedOrder =
+            Order.Data({ id: 0, payload: payload, settlementTimestamp: uint248(block.timestamp) });
 
         vm.expectEmit({ emitter: address(perpsEngine) });
         emit LogCreateOrder(users.naruto, perpsAccountId, ETH_USD_MARKET_ID, expectedOrder);
@@ -57,7 +58,8 @@ contract CreateOrder_Integration_Test is Base_Integration_Shared_Test {
 
         perpsEngine.createOrder({ payload: payload });
 
-        Order.Data memory expectedOrder = Order.Data({ id: 1, payload: payload, settlementTimestamp: block.timestamp });
+        Order.Data memory expectedOrder =
+            Order.Data({ id: 1, payload: payload, settlementTimestamp: uint248(block.timestamp) });
 
         vm.expectEmit({ emitter: address(perpsEngine) });
         emit LogCreateOrder(users.naruto, perpsAccountId, ETH_USD_MARKET_ID, expectedOrder);
